requirements_test.txt
pip install blackp-r requirements_dev.txt
pytest-homeassistant-custom-component==0.13.172
pytest-cov
